requirements.txt
paddleocr
python-docx
pymupdf
pdf2image
extract-msg
opencv-python
matplotlib
